@@ -1,20 +1,28 @@
 {
-	"name": "suin/php-rss-writer",
-	"type": "library",
-	"description": "Yet another simple RSS writer library for PHP 5.3 or later.",
-	"keywords": ["rss", "generator", "writer", "feed", "php"],
-	"homepage": "https://github.com/suin/php-rss-writer",
-	"license": "MIT",
-	"authors": [
-		{
-			"name": "Hidehito Nozawa aka Suin",
-			"email": "suinyeze@gmail.com"
-		}
-	],
-	"require": {
-		"php": ">=5.3.0"
-	},
-	"autoload": {
-		"psr-0": { "Suin\\RSSWriter": "src" }
-	}
+  "name": "suin/php-rss-writer",
+  "type": "library",
+  "description": "Yet another simple RSS writer library for PHP 5.4 or later.",
+  "keywords": [
+    "rss",
+    "generator",
+    "writer",
+    "feed",
+    "php"
+  ],
+  "homepage": "https://github.com/suin/php-rss-writer",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Hidehito Nozawa aka Suin",
+      "email": "suinyeze@gmail.com"
+    }
+  ],
+  "require": {
+    "php": ">=5.4.0"
+  },
+  "autoload": {
+    "psr-0": {
+      "Suin\\RSSWriter": "src"
+    }
+  }
 }
