@@ -1,8 +1,8 @@
 {
-	"require": {
-		"php":            ">=5.3.2",
-		"EHER/PHPUnit":   ">=1.6",
-		"suin/xoopsunit": ">=1.2",
-		"mockery/mockery": ">=0.7.2"
-	}
+  "require": {
+    "php": ">=5.4.0",
+    "EHER/PHPUnit": ">=1.6",
+    "suin/xoopsunit": ">=1.2",
+    "mockery/mockery": ">=0.7.2"
+  }
 }
