@@ -76,7 +76,7 @@ class ChannelTest extends \XoopsUnit\TestCase
         $item = $this->getMock($this->itemInterface);
         $channel = new Channel();
         $this->assertSame($channel, $channel->addItem($item));
-        $this->assertAttributeSame(array($item), 'items', $channel);
+        $this->assertAttributeSame([$item], 'items', $channel);
     }
 
     public function testAppendTo()
@@ -109,102 +109,102 @@ class ChannelTest extends \XoopsUnit\TestCase
         $now = time();
         $nowString = date(DATE_RSS, $now);
 
-        return array(
-            array(
+        return [
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                </channel>
+                ",
+                [
+                    'title'       => "GoUpstate.com News Headlines",
+                    'url'         => 'http://www.goupstate.com/',
                     'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
-                )
-            ),
-            array(
+                ]
+            ],
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-					<language>en-us</language>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                    <language>en-us</language>
+                </channel>
+                ",
+                [
+                    'title'       => "GoUpstate.com News Headlines",
+                    'url'         => 'http://www.goupstate.com/',
                     'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
-                    'language' => 'en-us',
-                )
-            ),
-            array(
+                    'language'    => 'en-us',
+                ]
+            ],
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-					<pubDate>{$nowString}</pubDate>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                    <pubDate>{$nowString}</pubDate>
+                </channel>
+                ",
+                [
+                    'title'       => "GoUpstate.com News Headlines",
+                    'url'         => 'http://www.goupstate.com/',
                     'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
-                    'pubDate' => $now,
-                )
-            ),
-            array(
+                    'pubDate'     => $now,
+                ]
+            ],
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-					<lastBuildDate>{$nowString}</lastBuildDate>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
-                    'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                    <lastBuildDate>{$nowString}</lastBuildDate>
+                </channel>
+                ",
+                [
+                    'title'         => "GoUpstate.com News Headlines",
+                    'url'           => 'http://www.goupstate.com/',
+                    'description'   => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
                     'lastBuildDate' => $now,
-                )
-            ),
-            array(
+                ]
+            ],
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-					<ttl>60</ttl>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                    <ttl>60</ttl>
+                </channel>
+                ",
+                [
+                    'title'       => "GoUpstate.com News Headlines",
+                    'url'         => 'http://www.goupstate.com/',
                     'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
-                    'ttl' => 60,
-                )
-            ),
-            array(
+                    'ttl'         => 60,
+                ]
+            ],
+            [
                 "
-				<channel>
-					<title>GoUpstate.com News Headlines</title>
-					<link>http://www.goupstate.com/</link>
-					<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-					<copyright>Copyright 2002, Spartanburg Herald-Journal</copyright>
-				</channel>
-				",
-                array(
-                    'title' => "GoUpstate.com News Headlines",
-                    'url' => 'http://www.goupstate.com/',
+                <channel>
+                    <title>GoUpstate.com News Headlines</title>
+                    <link>http://www.goupstate.com/</link>
+                    <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                    <copyright>Copyright 2002, Spartanburg Herald-Journal</copyright>
+                </channel>
+                ",
+                [
+                    'title'       => "GoUpstate.com News Headlines",
+                    'url'         => 'http://www.goupstate.com/',
                     'description' => "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.",
-                    'copyright' => "Copyright 2002, Spartanburg Herald-Journal",
-                )
-            ),
-        );
+                    'copyright'   => "Copyright 2002, Spartanburg Herald-Journal",
+                ]
+            ],
+        ];
     }
 
     public function testAppendTo_with_items()
@@ -226,24 +226,24 @@ class ChannelTest extends \XoopsUnit\TestCase
             ->attr('title', "GoUpstate.com News Headlines")
             ->attr('url', 'http://www.goupstate.com/')
             ->attr('description', "The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.")
-            ->attr('items', array($item1, $item2, $item3));
+            ->attr('items', [$item1, $item2, $item3]);
 
         $expect = '<?xml version="1.0" encoding="UTF-8" ?>
-			<channel>
-				<title>GoUpstate.com News Headlines</title>
-				<link>http://www.goupstate.com/</link>
-				<description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
-				<item>
-					<title>item1</title>
-				</item>
-				<item>
-					<title>item2</title>
-				</item>
-				<item>
-					<title>item3</title>
-				</item>
-			</channel>
-		';
+            <channel>
+                <title>GoUpstate.com News Headlines</title>
+                <link>http://www.goupstate.com/</link>
+                <description>The latest news from GoUpstate.com, a Spartanburg Herald-Journal Web site.</description>
+                <item>
+                    <title>item1</title>
+                </item>
+                <item>
+                    <title>item2</title>
+                </item>
+                <item>
+                    <title>item3</title>
+                </item>
+            </channel>
+        ';
 
         $this->assertXmlStringEqualsXmlString($expect, $channel->asXML()->asXML());
     }
